requirements/base.txt
# Django and extensions
Django==5.2.12
djangorestframework>=3.15,<4.0
django-cors-headers>=4.0,<5.0
django_celery_results>=2.5,<3.0
django-celery-beat>=2.5,<3.0
django-timezone-field>=6.0,<7.0

# Channels (WebSocket support)
channels>=4.0,<5.0
channels-redis>=4.0,<5.0
asgiref>=3.8,<4.0

# MongoDB
pymongo>=4.6,<5.0
mongoengine>=0.28,<1.0

# Supabase Storage
supabase>=2.0,<3.0

# HTTP clients
requests>=2.31,<3.0
aiohttp>=3.9,<4.0
websockets>=12.0,<13.0

# Celery
celery>=5.3,<6.0

# Utilities
python-dotenv>=1.0,<2.0
bcrypt>=4.1,<5.0
